+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.3 (unreleased)
 - Add `--version` flag (#1)
 - Add import/export from .env files (#2)
+- Fallback to environment variables when secret not in vault (#3)
 
 ## 0.1.2
 - Initial public release